refactor: separate auth modal logic
