+++ package.json
@@ -35,7 +35,8 @@
     "react-map-gl": "^7.1.7",
     "react-router-dom": "^6.23.0",
     "react-social-media-embed": "^2.5.13",
-    "yup": "^1.4.0"
+    "yup": "^1.4.0",
+    "zustand": "^5.0.4"
   },
   "devDependencies": {
     "@feature-sliced/eslint-config": "^0.1.1",
