--- src/entities/AuthModal/components/ContinueWithGoogle/ui/index.ts
@@ -1,1 +1,0 @@
-export { ContinueWithGoogleButton } from './ContinueWithGoogleButton';
