--- src/entities/AuthModal/components/SignUpWithEmail/ui/SignUpWithEmail.module.scss
@@ -1,58 +1,0 @@
-.content {
-  text-align: center;
-  padding: 0 20px;
-}
-
-.registerWithEmail {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.withGoogle {
-  padding-top: 20px;
-  text-align: center;
-}
-.or {
-  color: #d5d5d5;
-  font-weight: 400;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-  flex-shrink: 0;
-  padding: 20px 0 8px;
-  text-align: center;
-
-  &:before,
-  &::after {
-    content: '';
-    height: 1px;
-    width: 40px;
-    background: #d5d5d5;
-    margin-top: 2px;
-  }
-}
-
-.recaptcha {
-  font-size: 14px;
-  padding-bottom: 16px;
-  display: flex;
-  flex-direction: column;
-  color: #fd5353;
-  align-items: center;
-  justify-content: center;
-}
-
-.signIn {
-  padding-top: 20px;
-  font-size: 12px;
-  span {
-    cursor: pointer;
-    font-weight: 500;
-    &:hover {
-      border-bottom: 1px #303030 solid;
-    }
-  }
-}
