--- src/entities/AuthModal/index.ts
@@ -1,5 +1,0 @@
-export * from './ui';
-
-export { SignInWithEmail } from './components/SignInWithEmail';
-export { SignUpWithEmail } from './components/SignUpWithEmail';
-export { SuccessfulSignUp } from './components/SuccessfulSignUp';
