+++ src/features/ContinueWithGoogle/index.ts
@@ -1,0 +1,1 @@
+export { useWithGoogle } from './api/useWithGoogle';
