+++ src/features/SignInWithEmail/types/index.ts
@@ -1,0 +1,10 @@
+import type * as Yup from 'yup';
+import { type validationSchemaSignInWithEmail } from '../lib/validationSchema';
+
+export type SignInWithEmailData = Yup.InferType<typeof validationSchemaSignInWithEmail>;
+
+export interface SignInWithEmailProps {
+  hideAuthModal?: () => void;
+  onSwitchToSignUp: () => void;
+  continueWithSocial?: React.ReactNode[];
+}
