+++ src/features/SignUpWithEmail/types/index.ts
@@ -1,0 +1,10 @@
+import type * as Yup from 'yup';
+import { type validationSchemaSignUpWithEmail } from '../lib/validationSchema';
+
+export type SignUpWithEmailData = Yup.InferType<typeof validationSchemaSignUpWithEmail>;
+
+export interface SignUpWithEmailProps {
+  onSuccessfulSignUp: () => void;
+  onSwitchToSignIn: () => void;
+  continueWithSocial?: React.ReactNode[];
+}
