+++ src/pages/LoginPage/ui/LoginPage.tsx
@@ -1,13 +1,18 @@
 import { useEffect, useState } from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
-import { SignInWithEmail, SignUpWithEmail } from 'entities/AuthModal';
+import { useWithGoogle } from 'features/ContinueWithGoogle';
+import { SignInWithEmail } from 'features/SignInWithEmail';
+import { SignUpWithEmail } from 'features/SignUpWithEmail';
 import { useAuth } from 'shared/api';
 import { useAuthModal } from 'shared/context/Auth/AuthModalContext';
+
+import { GOOGLE_LOGIN_BUTTON_KEY, GoogleLoginButton } from 'shared/ui/GoogleLoginButton';
 import cls from './LoginPage.module.scss';
 
 export const LoginPage = () => {
   const { user } = useAuth();
   const { showSuccessfulSignUp } = useAuthModal();
+  const continueWithGoogle = useWithGoogle();
   const navigate = useNavigate();
   const location = useLocation();
 
@@ -29,6 +34,7 @@ export const LoginPage = () => {
             onSwitchToSignUp={() => {
               setIiSignIn(false);
             }}
+            continueWithSocial={[<GoogleLoginButton key={GOOGLE_LOGIN_BUTTON_KEY} onClick={continueWithGoogle} />]}
           />
         ) : (
           <SignUpWithEmail
@@ -36,6 +42,7 @@ export const LoginPage = () => {
             onSwitchToSignIn={() => {
               setIiSignIn(true);
             }}
+            continueWithSocial={[<GoogleLoginButton key={GOOGLE_LOGIN_BUTTON_KEY} onClick={continueWithGoogle} />]}
           />
         )}
       </div>
