+++ src/shared/stores/auth/constants/index.ts
@@ -1,0 +1,6 @@
+import { type AuthState } from '../types';
+
+export const INITIAL_STATE: AuthState = {
+  user: null,
+  isLoading: false,
+};
