+++ src/shared/stores/auth/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useAuthStore } from './useAuthStore';
