+++ src/shared/stores/auth/hooks/useAuthStore.ts
@@ -1,0 +1,6 @@
+import { create } from 'zustand';
+
+import { INITIAL_STATE } from '../constants';
+import { type AuthState } from '../types';
+
+export const useAuthStore = create<AuthState>(() => INITIAL_STATE);
