+++ src/shared/stores/auth/types/index.ts
@@ -1,0 +1,13 @@
+export interface User {
+  id: string;
+  displayName: string;
+  email: string;
+  avatar?: string;
+  createdAt?: Date;
+  isGoogleUserUserWithoutPassword: boolean;
+}
+
+export interface AuthState {
+  user: User | null;
+  isLoading: boolean;
+}
