+++ src/shared/ui/GoogleLoginButton/constants/index.ts
@@ -1,1 +1,3 @@
 export const GOOGLE_LOGIN_BUTTON_TEXT = 'Continue with Google';
+
+export const GOOGLE_LOGIN_BUTTON_KEY = 'google';
