+++ src/shared/ui/GoogleLoginButton/index.ts
@@ -1,1 +1,2 @@
 export { GoogleLoginButton } from './ui/GoogleLoginButton';
+export { GOOGLE_LOGIN_BUTTON_KEY } from './constants';
