+++ src/widgets/AuthModal/components/LoginRequired/index.ts
