+++ src/widgets/AuthModal/components/LoginRequired/types/index.ts
@@ -1,0 +1,4 @@
+export interface LoginRequiredProps {
+  onSwitchToSignIn: () => void;
+  onSwitchToSignUp: () => void;
+}
