+++ src/widgets/AuthModal/components/LoginRequired/ui/LoginRequired.module.scss
