+++ src/widgets/AuthModal/components/LoginRequired/ui/index.ts
