+++ src/widgets/AuthModal/components/SuccessfulSignUp/index.ts
