+++ src/widgets/AuthModal/components/SuccessfulSignUp/ui/SuccessfulSignUp.module.scss
