+++ src/widgets/AuthModal/components/SuccessfulSignUp/ui/SuccessfulSignUp.tsx
