+++ src/widgets/AuthModal/components/SuccessfulSignUp/ui/index.ts
