+++ src/widgets/AuthModal/ui/AuthModal.module.scss
