+++ src/widgets/Navbar/ui/Navbar.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { NavLink, useLocation } from 'react-router-dom';
 import { AuthIndicator } from 'features/AuthIndicator';
 import { SearchPlaces } from 'features/SearchPlaces';
-import { AuthModal } from 'entities/AuthModal';
+import { AuthModal } from 'widgets/AuthModal';
 import { Logo } from 'shared/ui/Logo';
 import cls from './Navbar.module.scss';
 export const Navbar = () => {
